+++ packages/jivs-engine/src/ValueHosts/ValidatableValueHostBase.ts
@@ -683,7 +683,20 @@ export abstract class ValidatableValueHostBase<TConfig extends ValidatableValueH
     }
         
     //#endregion validation results
-
+    /**
+     * Broadcasts the current state of the ValueHost to any listeners or managers that need to be aware of changes.
+     * FieldValueHost uses this to report its textvalue through ValueHostsManager.onTextValueChanged
+     * and validation state through ValueHostsManager.onValueHostValidationStateChanged.
+     * Normally those events are fired at appropriate times.
+     * However, when recreating ValueHostsManager with its state from a previous lifecycle,
+     * that state does not cause the usual events to be fired automatically.
+     * Calling broadcastState() ensures that the current state is communicated to all relevant listeners.
+     * This mostly targets pages generated on the server side, like MVC.
+     */
+    public override broadcastState(): void {
+        super.broadcastState();
+        this.invokeOnValueHostValidationStateChanged({});
+    }
 }
 
 /**