+++ packages/jivs-engine/src/ValueHosts/ValueHostBase.ts
@@ -332,7 +332,22 @@ export abstract class ValueHostBase<TConfig extends ValueHostConfig,
 
     //#endregion IValueHost
 
-    //#region State
+    //#region instancestate
+    /**
+     * Broadcasts the current state of the ValueHost to any listeners or managers that need to be aware of changes.
+     * FieldValueHost uses this to report its textvalue through ValueHostsManager.onTextValueChanged
+     * and validation state through ValueHostsManager.onValueHostValidationStateChanged.
+     * Normally those events are fired at appropriate times.
+     * However, when recreating ValueHostsManager with its state from a previous lifecycle,
+     * that state does not cause the usual events to be fired automatically.
+     * Calling broadcastState() ensures that the current state is communicated to all relevant listeners.
+     * This mostly targets pages generated on the server side, like MVC.
+     */
+    public broadcastState(): void
+    {
+
+    }
+    
     /* 
      * Current state for the associated ValueHost.
      * Only ValueHostsManager owns the state. This instance is a reference