+++ packages/jivs-engine/tests/ValueHosts/StaticValueHost.test.ts
@@ -237,6 +237,9 @@ describe('toIStaticValueHost function', () => {
         }
         _captureState(stateContainer: StateContainer): void {
             throw new Error("Method not implemented.");
+        }
+        broadcastState(): void {
+            throw new Error("Method not implemented.");
         }        
     }
     test('Passing object with interface match returns same object.', () => {