+++ packages/jivs-engine/tests/ValueHosts/ValidatorsValueHostBase.test.ts
@@ -3099,7 +3099,10 @@ describe('toIValidatorsValueHostBase function', () => {
         _captureState(stateContainer: StateContainer): void {
             throw new Error('Method not implemented.');
         }
-
+        broadcastState(): void
+        {
+            throw new Error("Method not implemented.");
+        }
     }
     test('Passing object with interface match returns same object.', () => {
         let testItem = new TestIValidatorsValueHostBaseImplementation();